collsion impact, improved enemy spawning, still a bug with enemys that are spawned after player.

+++ BuasGameSDL2/Enemy.h
@@ -12,7 +12,7 @@ public:
 	void onCollision(std::string otherTag, GameObject* other);
 	void SetEnemyPosScale(int beginYPos);
 	void AnimateFrame(int Anim);
-	float speed = 4;
+	int speed = 4;
 	float time;
 	float maxTime = 10;
 	float currentFrame;
